requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
audioread=2.1.9=py39hcbf5309_2
blas=2.106=mkl
blosc=1.21.1=hcbbf2c4_0
brotli=1.0.9=h8ffe710_6
brotli-bin=1.0.9=h8ffe710_6
bzip2=1.0.8=h8ffe710_4
c-blosc2=2.0.4=h09319c2_1
ca-certificates=2022.5.18.1=h5b45459_0
certifi=2022.5.18.1=py39hcbf5309_0
cffi=1.15.0=py39h0878f49_0
cfitsio=4.0.0=hd67004f_0
charls=2.2.0=h39d44d4_0
cloudpickle=2.1.0=pyhd8ed1ab_0
cudatoolkit=10.2.89=hb195166_9
cycler=0.11.0=pyhd8ed1ab_0
cytoolz=0.11.2=py39hb82d6ee_2
dask-core=2022.5.0=pyhd8ed1ab_0
decorator=5.1.1=pyhd8ed1ab_0
fonttools=4.28.5=py39hb82d6ee_0
freetype=2.10.4=h546665d_1
fsspec=2022.5.0=pyhd8ed1ab_0
future=0.18.2=py39hcbf5309_4
giflib=5.2.1=h8d14728_2
imagecodecs=2021.11.20=py39he391c9c_1
imageio=2.13.5=pyh239f2a4_0
intel-openmp=2022.0.0=h57928b3_3663
jbig=2.1=h8d14728_2003
joblib=1.1.0=pyhd8ed1ab_0
jpeg=9d=h8ffe710_0
jxrlib=1.1=h8ffe710_2
kiwisolver=1.3.2=py39h2e07f2f_1
krb5=1.19.3=hc8ab02b_0
lcms2=2.12=h2a16943_0
lerc=3.0=h0e60522_0
libaec=1.0.6=h39d44d4_0
libblas=3.9.0=6_mkl
libbrotlicommon=1.0.9=h8ffe710_6
libbrotlidec=1.0.9=h8ffe710_6
libbrotlienc=1.0.9=h8ffe710_6
libcblas=3.9.0=6_mkl
libcurl=7.83.1=h789b8ee_0
libdeflate=1.8=h8ffe710_0
libffi=3.4.2=h8ffe710_5
libflac=1.3.3=h0e60522_1
liblapack=3.9.0=6_mkl
liblapacke=3.9.0=6_mkl
libogg=1.3.4=h8ffe710_1
libopus=1.3.1=h8ffe710_1
libpng=1.6.37=h1d00b33_2
librosa=0.7.2=py_1
libsndfile=1.0.31=h0e60522_1
libssh2=1.10.0=h9a1e1f7_2
libtiff=4.3.0=hd413186_2
libuv=1.43.0=h8ffe710_0
libvorbis=1.3.7=h0e60522_0
libwebp-base=1.2.2=h8ffe710_1
libzlib=1.2.11=h8ffe710_1013
libzopfli=1.0.3=h0e60522_0
llvmlite=0.36.0=py39ha0cd8c8_0
locket=1.0.0=pyhd8ed1ab_0
lz4-c=1.9.3=h8ffe710_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
matplotlib-base=3.5.1=py39h581301d_0
mir_eval=0.6=pyh9f0ad1d_0
mkl=2020.4=hb70f87d_311
msys2-conda-epoch=20160418=1
munkres=1.1.4=pyh9f0ad1d_0
networkx=2.8.2=pyhd8ed1ab_0
numba=0.53.1=py39hb8cd55e_1
numpy=1.22.0=py39h6331f09_0
olefile=0.46=pyh9f0ad1d_1
opencv-python=4.5.5.62=pypi_0
openjpeg=2.4.0=hb211442_1
openssl=3.0.3=h8ffe710_0
packaging=21.3=pyhd8ed1ab_0
partd=1.2.0=pyhd8ed1ab_0
pillow=8.4.0=py39h916092e_0
pip=21.3.1=pyhd8ed1ab_0
portaudio=19.6.0=h0e60522_4
pycparser=2.21=pyhd8ed1ab_0
pyparsing=3.0.7=pyhd8ed1ab_0
pysoundfile=0.10.3.post1=pyhd3deb0d_0
python=3.9.9=hcf16a7b_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-sounddevice=0.4.1=pyh9f0ad1d_0
python_abi=3.9=2_cp39
pytorch=1.10.1=py3.9_cuda10.2_cudnn7_0
pytorch-mutex=1.0=cuda
pytube=11.0.2=pyhd8ed1ab_0
pywavelets=1.3.0=py39h5d4886f_1
pyyaml=6.0=py39hb82d6ee_4
resampy=0.2.2=py_0
scikit-image=0.19.2=py39h2e25243_0
scikit-learn=1.0.2=py39he931e04_0
scipy=1.7.3=py39hc0c34ad_0
setuptools=60.5.0=py39hcbf5309_0
six=1.16.0=pyh6c4a22f_0
snappy=1.1.9=h82413e6_1
sounddevice=0.4.4=pypi_0
sqlite=3.37.0=h8ffe710_0
threadpoolctl=3.0.0=pyh8a188c0_0
tifffile=2022.5.4=pyhd8ed1ab_0
tk=8.6.11=h8ffe710_1
toolz=0.11.2=pyhd8ed1ab_0
torchaudio=0.10.1=py39_cu102
torchvision=0.11.2=py39_cu102
typing_extensions=4.0.1=pyha770c72_0
tzdata=2021e=he74cb21_0
ucrt=10.0.20348.0=h57928b3_0
vc=14.2=hb210afc_6
vs2015_runtime=14.29.30037=h902a5da_6
wheel=0.37.1=pyhd8ed1ab_0
xz=5.2.5=h62dcd97_1
yaml=0.2.5=h8ffe710_2
zfp=0.5.5=h0e60522_8
zlib=1.2.11=h8ffe710_1013
zstd=1.5.1=h6255e5f_0
